requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
notion-client==2.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
openai==1.3.0
elevenlabs==0.2.26
ffmpeg-python==0.2.0
python-multipart==0.0.6
apscheduler==3.10.4
pydantic==2.4.2
requests==2.31.0
python-jose==3.3.0 